delete batch-brevcodes, move brevmetadata-dto to its own class
