--- server/src/main/java/no/nav/pensjon/testdata/consumer/brev/BatchBrevTypeCode.java
@@ -1,401 +1,0 @@
-package no.nav.pensjon.testdata.consumer.brev;
-
-public enum BatchBrevTypeCode {
-    AFPEO_BREV_A,
-    /**
-     * -
-     */
-    AFPEO_BREV_B,
-    /**
-     * -
-     */
-    AFPEO_BREV_C,
-    /**
-     * -
-     */
-    AFPEO_BREV_D,
-    /**
-     * -
-     */
-    AFPEO_TILBAKEKREVING,
-    /**
-     * -
-     */
-    AFPP_AVSL_AUTO,
-    /**
-     * -
-     */
-    AFPP_INNVILG_AUTO,
-    /**
-     * -
-     */
-    AFP_3MND_67,
-    /**
-     * -
-     */
-    AFP_PR_NYTT_VEDTAK,
-    /**
-     * -
-     */
-    ALDER_IVERK,
-    /**
-     * -
-     */
-    ALDER_SAMM_IVERK,
-    /**
-     * -
-     */
-    AO_AP_AFP_67_2011,
-    /**
-     * -
-     */
-    AO_AP_BT_18_2011,
-    /**
-     * -
-     */
-    AO_AP_ET_2011,
-    /**
-     * -
-     */
-    AO_AP_ET_60_2011,
-    /**
-     * -
-     */
-    AO_AP_ET_BT_2011,
-    /**
-     * Vedtak - endring av alderspensjon ved aldersovergang
-     */
-    AO_AP_FT_1967,
-    /**
-     * -
-     */
-    AO_AP_FAP_PL_67_2011,
-    /**
-     * -
-     */
-    AO_AP_FAP_PL_67_2016,
-    /**
-     * -
-     */
-    AO_AP_GJP_FULL_67,
-    /**
-     * -
-     */
-    AO_AP_GRAD_AP_75,
-    /**
-     * -
-     */
-    AO_AP_UP_67_2011,
-    /**
-     * -
-     */
-    AP_AVSLAG_AUTO,
-    /**
-     * -
-     */
-    AP_AVSL_UTL_AUTO,
-    /**
-     * -
-     */
-    AP_ENDR_GRAD_AUTO,
-    /**
-     * -
-     */
-    AP_FULLTT_BELOPENDR,
-    /**
-     * -
-     */
-    AP_INNVILG_AUTO,
-    /**
-     * -
-     */
-    AP_INNVILG_UTL,
-    /**
-     * -
-     */
-    AP_INNVILG_UTL_AUTO,
-    /**
-     * -
-     */
-    AP_NYTT_VEDTAK,
-    /**
-     * -
-     */
-    AUTO_AP_ENDR_RETTING,
-    /**
-     * -
-     */
-    AUTO_INFOBREV_BP,
-    /**
-     * -
-     */
-    AUTO_INFOBREV_GJENL,
-    /**
-     * -
-     */
-    BP_UTBET_MINKER,
-    /**
-     * -
-     */
-    BP_UTBET_OKER,
-    /**
-     * -
-     */
-    DUMMYBREV,
-    /**
-     * -
-     */
-    ENDRING_UT,
-    /**
-     * -
-     */
-    ENDR_UT_PGA_INNT,
-    /**
-     * -
-     */
-    FAM_PL_3MND_67,
-    /**
-     * -
-     */
-    FAM_PL_INFO_67,
-    /**
-     * -
-     */
-    FORSORG_BT_OPPH,
-    /**
-     * -
-     */
-    FORSORG_ET_BT_OPPH,
-    /**
-     * -
-     */
-    FORSORG_ET_OPPH,
-    /**
-     * -
-     */
-    FYLLER_67,
-    /**
-     * -
-     */
-    FYLLER_67_INGEN_YT,
-    /**
-     * -
-     */
-    GJP_AVKORT_3MND_67,
-    /**
-     * -
-     */
-    GJP_AVKORT_INFO_67,
-    /**
-     * -
-     */
-    GJP_FULL_3MND_67,
-    /**
-     * -
-     */
-    GJP_FULL_INFO_67,
-    /**
-     * -
-     */
-    GJR_UTBET_MINKER,
-    /**
-     * -
-     */
-    GJR_UTBET_OKER,
-    /**
-     * -
-     */
-    INFOBREV_VP_ET,
-    /**
-     * -
-     */
-    KONV_UP_TIL_UT,
-    /**
-     * -
-     */
-    KONV_YT_ALDERSOVERG,
-    /**
-     * -
-     */
-    LEVEATT,
-    /**
-     * -
-     */
-    LEVEATT_ALT,
-    /**
-     * -
-     */
-    LEVEATT_PURRING,
-    /**
-     * -
-     */
-    LEVEATT_PURRING_ALT,
-    /**
-     * -
-     */
-    NY_BER_UT_BT,
-    /**
-     * -
-     */
-    NY_BER_UT_ENDR_OPPTJ,
-    /**
-     * -
-     */
-    OMREGN_UP_TIL_UT,
-    /**
-     * -
-     */
-    OMR_BP_OPPH,
-    /**
-     * -
-     */
-    OMSORGP_BARN,
-    /**
-     * -
-     */
-    OMSORGP_EGENMLD,
-    /**
-     * -
-     */
-    OMSORGP_GODSKRIVING,
-    /**
-     * -
-     */
-    OMSORGP_INNVILG,
-    /**
-     * -
-     */
-    OMSORGP_ORIENT_GODS,
-    /**
-     * -
-     */
-    OMSORGP_SOKNAD_FAR,
-    /**
-     * -
-     */
-    OMSORGP_SOKNAD_MOR,
-    /**
-     * -
-     */
-    OO_AVGIVER,
-    /**
-     * -
-     */
-    OO_MOTTAKER,
-    /**
-     * -
-     */
-    OPPHOR_ET_GG,
-    /**
-     * -
-     */
-    OPPHOR_ET_NN,
-    /**
-     * -
-     */
-    P_UTBETALING_MINKER,
-    /**
-     * -
-     */
-    P_UTBETALING_OKER,
-    /**
-     * -
-     */
-    REV_TIL_AFP_PGA_DOD,
-    /**
-     * -
-     */
-    REV_TIL_AP_PGA_DOD,
-    /**
-     * -
-     */
-    REV_TIL_UP_PGA_DOD,
-    /**
-     * -
-     */
-    SAMMENSTOT_AP,
-    /**
-     * -
-     */
-    TILST_AFP,
-    /**
-     * -
-     */
-    TILST_AP,
-    /**
-     * -
-     */
-    TILST_GJP,
-    /**
-     * -
-     */
-    TILST_OMRG_AP,
-    /**
-     * -
-     */
-    TILST_UP,
-    /**
-     * -
-     */
-    UP_FULLTT_BELOPENDR,
-    /**
-     * -
-     */
-    UP_GRAD_AP_67,
-    /**
-     * -
-     */
-    UP_GRAD_IKKE_AP_67,
-    /**
-     * -
-     */
-    UP_HEL_67,
-
-    UT_EO_VARSEL_FU,
-
-    UT_EO_VEDTAK_EB,
-    /**
-     * -
-     */
-    VARSEL_TILB_KREV,
-    /**
-     * -
-     */
-    //VARS_LANG_SAKS_AUTO,
-    /**
-     * -
-     */
-    VEDTAK_SKJERMT,
-    /**
-     * -
-     */
-    VEDTAK_TILB_KREV,
-    TILST_DOD_UT,
-    ENDRING_UT_BT,
-    /**
-     * Vedtak - opphør av barnetillegg (automatisk)
-     */
-    OPPHOR_ENDRING_UT_BT,
-
-
-    AUTO_VTK_EPS_DOD_AP,
-    AUTO_VTK_EPS_DOD_AFP,
-    /**
-     * Vedtak - omregning av uføretrygd til alderspensjon
-     */
-    AO_AP_UT_67_2016,
-    /**
-     * Varsel - endring av uføretrygd på grunn av inntekt
-     */
-    INFO_INNT_ENDR_UT,
-    /**
-     * Informasjon om gjenlevendetillegg i alderspensjonen
-     */
-    GJP_INFOBREV_AP_GJT,
-    /**
-     * Automatisk brev engangsbruk
-     */
-    ADHOC_INFOBREV_AUTO,
-
-    AP_INNV_AO_UT_AUTO;
-}
