+++ server/src/main/java/no/nav/pensjon/testdata/controller/BrevController.java
@@ -8,6 +8,7 @@ import io.swagger.annotations.Tag;
 import no.nav.pensjon.testdata.configuration.support.JdbcTemplateWrapper;
 import no.nav.pensjon.testdata.consumer.brev.BrevConsumer;
 import no.nav.pensjon.testdata.consumer.brev.BrevMetaDataConsumer;
+import no.nav.pensjon.testdata.consumer.brev.BrevMetadata;
 import no.nav.pensjon.testdata.controller.support.BestillBrevRequest;
 import no.nav.tjeneste.domene.pensjon.vedtaksbrev.binding.BestillAutomatiskBrevAdresseMangler;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -56,7 +57,7 @@ public class BrevController {
     }
 
     @GetMapping
-    public ResponseEntity<List<BrevMetaDataConsumer.Metadata>> hentBrevkoder(){
+    public ResponseEntity<List<BrevMetadata>> hentBrevkoder(){
         return ResponseEntity.ok(brevMetaData.getAllBrev());
     }
